moved query and mutation hooks to hooks/api and renamed them. created new componens for the tasks list. ran linter with fix option
