+++ src/screens/Home/TasksList/index.tsx
@@ -1,23 +1,20 @@
-import { formatDistance } from 'date-fns';
-import { es } from 'date-fns/locale';
 import { useState } from 'react';
-import { Text, View, Pressable, ViewProps } from 'react-native';
+import { Text, View } from 'react-native';
 
-import { useTasksListQuery } from './queries';
+import Header from './Header';
+import Item from './Item';
 
 import { TaskStatus } from '@/api/graphql';
-import { Badge, BadgeText } from '@/components/ui/badge';
 import { Button, ButtonText } from '@/components/ui/button';
+import { useGetMyAssignedTasks } from '@/hooks/api/tasks/useGetMyAssignedTasks';
 import { cn } from '@/lib/utils';
 import { HomeTabScreenProp } from '@/navigation/types';
 
-type TaskItem = NonNullable<Props['tasksQuery']['data']>['myAssignedTasks'][0];
-
-type Props = Pick<HomeTabScreenProp, 'navigation'> & {
-    tasksQuery: ReturnType<typeof useTasksListQuery>;
+export type TaskListProps = Pick<HomeTabScreenProp, 'navigation'> & {
+    tasksQuery: ReturnType<typeof useGetMyAssignedTasks>;
 };
 
-const TasksList = ({ navigation, tasksQuery }: Props) => {
+const TasksList = ({ navigation, tasksQuery }: TaskListProps) => {
     const [taskStatus, setTaskStatus] = useState<TaskStatus>(TaskStatus.Pendiente);
 
     const filteredTasks = tasksQuery.data?.myAssignedTasks.filter((task) => {
@@ -115,78 +112,6 @@ const TasksList = ({ navigation, tasksQuery }: Props) => {
         <View className="my-4">
             <Header title="Cargando..." />
         </View>
-    );
-};
-
-type HeaderProps = {
-    title: string;
-    description?: string;
-};
-
-const Header = ({ title, description }: HeaderProps) => {
-    return (
-        <View className="flex flex-column items-center justify-between px-4 mb-4">
-            <Text className="font-bold text-lg">{title}</Text>
-            <Text>{description}</Text>
-        </View>
-    );
-};
-
-type ItemProps = {
-    task: TaskItem;
-    navigation: Props['navigation'];
-    style?: ViewProps['style'];
-};
-
-const Item = ({ task, navigation, style }: ItemProps) => {
-    const branch = task.branch;
-    const client = branch.client;
-
-    const [isPressed, setIsPressed] = useState(false);
-
-    return (
-        <Pressable
-            className={cn(
-                'flex flex-col items-start rounded-lg border border-border p-3 text-left text-sm transition-all',
-                isPressed && 'bg-accent',
-            )}
-            onPress={() => {
-                navigation.navigate('Task', {
-                    id: task.id,
-                });
-            }}
-            onPressIn={() => {
-                setIsPressed(true);
-            }}
-            onPressOut={() => {
-                setIsPressed(false);
-            }}
-            style={style}
-        >
-            <View className="flex w-full flex-col mb-2">
-                <View className="flex flex-row items-center mb-1">
-                    <View className="flex flex-row items-center">
-                        <Text className="font-semibold">{task.business.name}</Text>
-                    </View>
-                    <Text className={cn('ml-auto text-xs', 'text-muted-foreground')}>
-                        hace{' '}
-                        {formatDistance(task.createdAt, new Date(), {
-                            locale: es,
-                        })}
-                    </Text>
-                </View>
-
-                <Text className="text-xs font-medium">{client.name}</Text>
-            </View>
-
-            <Text numberOfLines={2} className="text-xs text-muted-foreground mb-2">
-                {task.description}
-            </Text>
-
-            <Badge>
-                <BadgeText>{task.taskType}</BadgeText>
-            </Badge>
-        </Pressable>
     );
 };
